Move router to HTTP too
